+++ apps/typesync/client/src/clients/graphql.ts
@@ -1,4 +1,0 @@
-import { Graph } from '@graphprotocol/grc-20';
-import { GraphQLClient } from 'graphql-request';
-
-export const graphqlClient = new GraphQLClient(`${Graph.TESTNET_API_ORIGIN}/graphql`);