+++ packages/hypergraph/src/index.ts
@@ -1,4 +1,4 @@
-export * as Typesync from './cli/services/Model.js';
+export { Id } from '@graphprotocol/grc-20';
 export * as Connect from './connect/index.js';
 export * as Entity from './entity/index.js';
 export * as Identity from './identity/index.js';