+++ packages/hypergraph/test/mapping/Mapping.test.ts
@@ -12,11 +12,11 @@ import {
 describe('Mapping', () => {
   describe('mapSchemaDataTypeToGRC20PropDataType', () => {
     it('should be able to map the schema dataType to the correct GRC-20 dataType', () => {
-      expect(mapSchemaDataTypeToGRC20PropDataType('Boolean')).toEqual('CHECKBOX');
+      expect(mapSchemaDataTypeToGRC20PropDataType('Boolean')).toEqual('BOOLEAN');
       expect(mapSchemaDataTypeToGRC20PropDataType('Number')).toEqual('NUMBER');
       expect(mapSchemaDataTypeToGRC20PropDataType('Date')).toEqual('TIME');
       expect(mapSchemaDataTypeToGRC20PropDataType('Point')).toEqual('POINT');
-      expect(mapSchemaDataTypeToGRC20PropDataType('String')).toEqual('TEXT');
+      expect(mapSchemaDataTypeToGRC20PropDataType('String')).toEqual('STRING');
       expect(mapSchemaDataTypeToGRC20PropDataType('Relation(Event)')).toEqual('RELATION');
     });
   });