+++ packages/typesync/src/index.ts
@@ -1,2 +1,0 @@
-export * as Mapping from './Mapping.js';
-export * as Utils from './Utils.js';