+++ vitest.config.ts
@@ -2,6 +2,6 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   test: {
-    projects: ['./packages/*', './apps/events', './apps/typesync', './apps/create-hypergraph'],
+    projects: ['./packages/*', './apps/events', './apps/create-hypergraph'],
   },
 });