+++ apps/server/CHANGELOG.md
@@ -1,5 +1,24 @@
 # server
 
+## 0.1.5
+### Patch Changes
+
+- Updated dependencies [e324e68]
+  - @graphprotocol/hypergraph@0.4.1
+
+## 0.1.4
+### Patch Changes
+
+- Updated dependencies [9b29006]
+  - @graphprotocol/hypergraph@0.4.0
+
+## 0.1.3
+### Patch Changes
+
+- Updated dependencies [b37483e]
+- Updated dependencies [f4fd295]
+  - @graphprotocol/hypergraph@0.3.1
+
 ## 0.1.2
 ### Patch Changes